+++ Engine/UI/Panel/PanelScene.h
@@ -4,8 +4,6 @@
 #include "Panel.h"
 #include <MathGeoLib.h>
 
-class Billboard;
-
 class PanelScene : public Panel
 {
 public:
@@ -13,7 +11,6 @@ public:
 	~PanelScene();
 
 	void Render() override;
-	void RenderDebugDraws();
 
 	void MousePicking(const float2& mouse_position);
 	void SceneDropTarget();
@@ -22,12 +19,6 @@ private:
 	void RenderSceneBar();
 
 	void RenderEditorDraws();
-
-	void RenderCameraFrustum() const;
-	void RenderOutline() const;
-	void RenderBoundingBoxes() const;
-	void RenderGlobalBoundingBoxes() const;
-	void RenderBillboards() const;
 
 	void RenderGizmo();
 	void RenderSceneCameraGizmo() const;
@@ -40,9 +31,6 @@ private:
 	float scene_window_content_area_height = 0;
 
 	float2 scene_window_content_area_pos = float2::zero;
-
-	Billboard* camera_billboard = nullptr;
-	Billboard* light_billboard = nullptr;
 };
 
 #endif //_PANELSCENE_H_
